cloud_functions/pg_reference_extractor/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-secret-manager==2.*
paramiko==3.*
psycopg2-binary==2.*
sshtunnel==0.4.*
